+++ packages/bridge-controller/src/index.ts
@@ -228,6 +228,7 @@ export {
   formatChainIdToHex,
   formatAddressToCaipReference,
   formatAddressToAssetId,
+  formatChainIdToDec,
 } from './utils/caip-formatters.js';
 
 export { extractTradeData } from './utils/trade-utils.js';